fix contract deployments not showing up (#2759)
* fix contract deployments not showing up

I'm really not happy with this.

Since the Signer consists of many, deeply nested, components, which in many cases
just pass props through, it's hard to trace this. `<TransactionPending>` is
supposed to fetch data from Parity (smart component) and pass it on to its (dumb)
children. For that, it needs to know implementation details of them.

* eslint :lipstick: